Review 1 Issue 2 : fix lateinit to nullable for Activity and change fragment's bindin to nullable + backing field
