+++ search/src/main/java/ru/practicum/android/diploma/search/presentation/ui/CountryFragment.kt
@@ -27,4 +27,9 @@ class CountryFragment : Fragment() {
             findNavController().navigateUp()
         }
     }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
 }
