+++ search/src/main/java/ru/practicum/android/diploma/search/presentation/ui/FilterFragment.kt
@@ -34,4 +34,9 @@ class FilterFragment : Fragment() {
             findNavController().navigateUp()
         }
     }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
 }
