+++ search/src/main/java/ru/practicum/android/diploma/search/presentation/ui/PlaceFragment.kt
@@ -34,4 +34,9 @@ class PlaceFragment : Fragment() {
             findNavController().navigateUp()
         }
     }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
+    }
 }
