--- search/src/main/java/ru/practicum/android/diploma/search/presentation/ui/VacancyFragment.kt
@@ -1,30 +1,0 @@
-package ru.practicum.android.diploma.search.presentation.ui
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.navigation.fragment.findNavController
-import ru.practicum.android.diploma.search.databinding.FragmentVacancyBinding
-
-class VacancyFragment : Fragment() {
-    private var _binding: FragmentVacancyBinding? = null
-    private val binding get() = _binding!!
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        _binding = FragmentVacancyBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        binding.buttonLeftVacancy.setOnClickListener {
-            findNavController().navigateUp()
-        }
-    }
-}
