+++ team/src/main/java/ru/practicum/android/diploma/team/presentation/ui/TeamFragment.kt
@@ -5,17 +5,25 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import ru.practicum.android.diploma.team.R
+import ru.practicum.android.diploma.team.databinding.FragmentTeamBinding
 
 class TeamFragment : Fragment() {
 
+    private var _binding: FragmentTeamBinding? = null
+    private val binding get() = _binding!!
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_team, container, false)
+        _binding = FragmentTeamBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        _binding = null
     }
 
 }
